Create README.md
# postit
My very first commit at all. I tried to focus and implement importante functions. It is an web application project where you can post it new messages so 
users can read it like regular post it's.

It is an web application with CRUD using PYTHON 3.2 as main language, HTML and CSS for the pages, a little of JS and FLASK as framework.

It is quite simple and easy to use the application. So far, we have basic functions as Register, Login, Create, Delete and Edit messages.

It was not easy to keep a decent front end job because it was my first experience exploring HTML and CSS but I tried to focus on decent look
and responsivity.

As back end choices, I kept all my attention with important functions, data safety, and tried to protect all ways in by url or bugs.

on MySql we have only 2 tables:
<table>
  <tr>
    <td>USERS:</td>
    <td>ID:</td>
    <td>Name:</td>
    <td>Username:</td>
    <td>Email:</td>
    <td>Password:</td>
    <td>Register_date</td>
  </tr>
    <td>MESSAGES:</td>
    <td>ID:</td>
    <td>Author:</td>
    <td>Create_date:</td>
</table>

Passfields are encrypted with sha256 library.
<table>
  <td>Techs:</td>
  <td>Python 3*</td>
  <td>Flask</td>
  <td>HTML</td>
  <td>CSS</td>
  <td>JS</td>
</table>
How to run:
config mysql
run app.py

This is a going on work where will have some changes and additions along the way.